15 warnings

+++ mettagrid/src/metta/mettagrid/actions/attack_nearest.hpp
@@ -20,7 +20,7 @@ protected:
   bool _handle_action(Agent* actor, [[maybe_unused]] ActionArg arg) override {
     // Scan the space to find the nearest agent. Prefer the middle (offset 0) before the edges (offset -1, 1).
     for (short distance = 1; distance < 4; distance++) {
-      for (int i = 0; i < 3; i++) {
+      for (short i = 0; i < 3; i++) {
         short offset = i;
         if (offset == 2) {
           // Sort of a mod 3 operation.
